Enrich FileInfo with cpio header values
Make the FileInfor more complete to allow transformations to e.g. tar
archives where additional info like GID and UID is needed.

Since the cpio headers are only available while streaming the archive,
the FileInfo entries are enriched on the fly.

The RPM header tags don't export GID and UID, therefor the CPIO header
seems to be needed for this info.

Signed-off-by: Roman Mohr <rmohr@redhat.com>